CsvHelper dll Removal
Removed references to a thrid party CsvHelper dll and implented the same using string split in program itself.
