+++ csvscan.csproj.user
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
-    <StartArguments>-v -r "C:\Projects\Mass Data CSV\sess 001" "C:\Projects\Mass Data CSV\filters.txt" "C:\Projects\Mass Data CSV\Output"</StartArguments>
+    <StartArguments>-v -r "C:\Projects\Mass Data CSV\sess 003" "C:\Projects\Mass Data CSV\filters.txt" "C:\Projects\Mass Data CSV\Output"</StartArguments>
   </PropertyGroup>
   <PropertyGroup>
     <PublishUrlHistory />
@@ -12,6 +12,7 @@
     <ErrorReportUrlHistory />
     <FallbackCulture>en-US</FallbackCulture>
     <VerifyUploadedFiles>false</VerifyUploadedFiles>
+    <ProjectView>ProjectFiles</ProjectView>
   </PropertyGroup>
   <PropertyGroup>
     <EnableSecurityDebugging>false</EnableSecurityDebugging>
